refactor dbt test logic into dedicated file
